requirements.txt
aws-cdk.core
aws-cdk.aws-cloudformation
aws-cdk.aws-lambda
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_rds
aws-cdk.aws_apigateway
